refactor!: merge proc-macros
