--- awg-proc-macro/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "actix-grants-proc-macro"
-version = "2.0.2"
-description = "A proc-macro way to validate user permissions for `actix-web-grants` crate."
-keywords = ["actix", "security", "grants", "codegen", "proc_macro"]
-readme = "../actix-web-grants/README.md"
-authors.workspace = true
-repository.workspace = true
-homepage.workspace = true
-categories.workspace = true
-license.workspace = true
-edition.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1"
-proc-macro2 = "1.0"
-syn = { version = "1.0", features = ["full", "derive", "extra-traits"] }
-
-[dev-dependencies]
-actix-web = { version = "4.3", default-features = false }
-actix-web-grants =  { workspace = true }
