+++ poem-grants/Cargo.toml
@@ -17,11 +17,11 @@ path = "src/lib.rs"
 
 [features]
 default = ["macro-check"]
-macro-check = ["poem-grants-proc-macro"]
+macro-check = ["protect-endpoints-proc-macro"]
 
 [dependencies]
 poem = "1"
-poem-grants-proc-macro = { workspace = true, optional = true }
+protect-endpoints-proc-macro = { workspace = true, features = ["rocket"], optional = true }
 thiserror = "1"
 
 [dev-dependencies]
