--- poem-proc-macro/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "poem-grants-proc-macro"
-version = "1.0.0-beta.3"
-description = "A proc-macro way to validate user permissions for `poem-grants` crate."
-keywords = ["poem", "security", "grants", "codegen", "proc_macro"]
-readme = "../poem-grants/README.md"
-authors.workspace = true
-repository.workspace = true
-homepage.workspace = true
-categories.workspace = true
-license.workspace = true
-edition.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = "1.0"
-quote = "1"
-syn = { version = "1.0", features = ["full", "derive", "extra-traits"] }
-
-[dev-dependencies]
-poem = "1"
-poem-openapi = "3.0.6"
-poem-grants =  { workspace = true }
