+++ proc-macro/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "rocket-grants-proc-macro"
-version = "0.1.1"
-description = "A proc-macro way to validate user permissions for `rocket-grants` crate."
-readme = "../rocket-grants/README.md"
+name = "protect-endpoints-proc-macro"
+version = "0.1.0"
+description = "A proc-macro way to protect your endpoints"
+readme = "../README.md"
 keywords = ["rocket", "authz", "security", "grants", "permissions"]
 authors.workspace = true
 repository.workspace = true
@@ -14,6 +14,11 @@ edition.workspace = true
 [lib]
 proc-macro = true
 
+[features]
+rocket = []
+actix-web = []
+poem = []
+
 [dependencies]
 darling = "0.20.3"
 proc-macro2 = "1.0"
@@ -21,6 +26,5 @@ quote = "1"
 syn = { version = "2.0", features = ["full", "derive", "extra-traits"] }
 
 [dev-dependencies]
-rocket-grants =  { workspace = true }
 rocket = { version = "0.5.0", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
