+++ proc-macro/src/expand.rs
@@ -1,10 +1,22 @@
 use darling::ast::NestedMeta;
-use proc_macro2::{Ident, Span, TokenStream as TokenStream2};
-use quote::{quote, ToTokens};
+use proc_macro2::{Ident, TokenStream as TokenStream2};
+use quote::quote;
 use std::ops::Deref;
-use syn::{ItemFn, Meta, ReturnType};
+use syn::{Block, ItemFn, Meta};
 
-const AUTH_DETAILS: &str = "_auth_details_";
+#[cfg(feature = "actix-web")]
+mod actix_web;
+#[cfg(feature = "poem")]
+mod poem;
+#[cfg(feature = "rocket")]
+mod rocket;
+
+#[derive(Clone)]
+pub(crate) enum FnType {
+    Fn(ItemFn),
+    #[cfg(feature = "poem")]
+    Method(syn::ImplItemFn),
+}
 
 #[derive(Debug)]
 enum Condition {
@@ -21,73 +33,22 @@ struct Conditions(Vec<Condition>);
 pub(crate) struct Args {
     cond: Condition,
     ty: Option<syn::Expr>,
+    error_fn: Option<Ident>,
 }
 
 pub(crate) struct ProtectEndpoint {
-    func: ItemFn,
+    func: FnType,
     args: Args,
 }
 
 impl ProtectEndpoint {
-    pub fn new(args: Args, func: ItemFn) -> syn::Result<Self> {
-        Ok(Self { func, args })
-    }
-}
-
-impl ToTokens for ProtectEndpoint {
-    fn to_tokens(&self, output: &mut TokenStream2) {
-        let func_vis = &self.func.vis;
-        let func_block = &self.func.block;
-
-        let fn_sig = &self.func.sig;
-        let fn_attrs = &self.func.attrs;
-        let fn_name = &fn_sig.ident;
-        let fn_generics = &fn_sig.generics;
-        let fn_args = &fn_sig.inputs;
-        let fn_async = &fn_sig.asyncness.unwrap();
-        let fn_output = match &fn_sig.output {
-            ReturnType::Type(ref _arrow, ref ty) => ty.to_token_stream(),
-            ReturnType::Default => {
-                quote! {()}
-            }
-        };
-
-        let condition = self.args.cond.to_tokens(self.args.ty.is_some());
-
-        let ty = self
-            .args
-            .ty
-            .as_ref()
-            .map(syn::Expr::to_token_stream)
-            .unwrap_or(quote! {String});
-
-        let condition = quote!(if #condition);
-        let auth_details: Ident = Ident::new(AUTH_DETAILS, Span::call_site());
-
-        let stream = quote! {
-            #(#fn_attrs)*
-            #func_vis #fn_async fn #fn_name #fn_generics(
-                #auth_details: rocket_grants::authorities::AuthDetails<#ty>,
-                #fn_args
-            ) -> Result<#fn_output, rocket::http::Status> {
-                use rocket_grants::authorities::AuthoritiesCheck;
-                #condition {
-                    let f = || async move #func_block;
-                    Ok(f().await)
-                } else {
-                    Err(rocket::http::Status::Forbidden)
-                }
-            }
-        };
-
-        output.extend(stream);
+    pub fn new(args: Args, func: FnType) -> Self {
+        Self { func, args }
     }
 }
 
 impl Condition {
-    fn to_tokens(&self, is_typed: bool) -> TokenStream2 {
-        let auth_details: Ident = Ident::new(AUTH_DETAILS, Span::call_site());
-
+    fn to_tokens(&self, auth_details: &Ident, is_typed: bool) -> TokenStream2 {
         match self {
             Condition::Any(nested) if nested.iter().all(Condition::is_value) => {
                 let vals = nested.iter().map(|c| match c {
@@ -120,12 +81,18 @@ impl Condition {
                 }
             }
             Condition::Any(nested) => {
-                let exprs: Vec<_> = nested.iter().map(|c| c.to_tokens(is_typed)).collect();
+                let exprs: Vec<_> = nested
+                    .iter()
+                    .map(|c| c.to_tokens(auth_details, is_typed))
+                    .collect();
 
                 quote! { #(#exprs)||* }
             }
             Condition::All(nested) => {
-                let exprs: Vec<_> = nested.iter().map(|c| c.to_tokens(is_typed)).collect();
+                let exprs: Vec<_> = nested
+                    .iter()
+                    .map(|c| c.to_tokens(auth_details, is_typed))
+                    .collect();
 
                 quote! { #(#exprs)&&* }
             }
@@ -199,6 +166,7 @@ impl darling::FromMeta for Args {
     fn from_list(items: &[NestedMeta]) -> darling::Result<Self> {
         let mut conditions = Vec::new();
         let mut ty = None;
+        let mut error_fn = None;
 
         let mut errors = ::darling::Error::accumulator();
 
@@ -210,6 +178,12 @@ impl darling::FromMeta for Args {
                             errors.push(darling::Error::duplicate_field("ty"));
                         } else {
                             ty = errors.handle(darling::FromMeta::from_expr(value));
+                        }
+                    } else if path.is_ident("error") {
+                        if ty.is_some() {
+                            errors.push(darling::Error::duplicate_field("error"));
+                        } else {
+                            error_fn = errors.handle(darling::FromMeta::from_expr(value));
                         }
                     } else if path.is_ident("expr") {
                         let cond = errors
@@ -244,9 +218,7 @@ impl darling::FromMeta for Args {
         }
 
         if conditions.is_empty() {
-            errors.push(darling::Error::custom(
-                "At least one condition must be specified",
-            ));
+            errors.push(darling::Error::custom("At least one condition must be specified"));
         }
 
         errors.finish()?;
@@ -257,7 +229,7 @@ impl darling::FromMeta for Args {
             Condition::All(Conditions(conditions))
         };
 
-        Ok(Args { cond, ty })
+        Ok(Args { cond, ty, error_fn })
     }
 }
 
@@ -266,5 +238,39 @@ impl Deref for Conditions {
 
     fn deref(&self) -> &Self::Target {
         &self.0
+    }
+}
+
+impl FnType {
+    fn sig(&self) -> &syn::Signature {
+        match self {
+            FnType::Fn(fun) => &fun.sig,
+            #[cfg(feature = "poem")]
+            FnType::Method(fun) => &fun.sig,
+        }
+    }
+
+    fn vis(&self) -> &syn::Visibility {
+        match self {
+            FnType::Fn(fun) => &fun.vis,
+            #[cfg(feature = "poem")]
+            FnType::Method(fun) => &fun.vis,
+        }
+    }
+
+    fn block(&self) -> &Block {
+        match self {
+            FnType::Fn(fun) => &fun.block,
+            #[cfg(feature = "poem")]
+            FnType::Method(fun) => &fun.block,
+        }
+    }
+
+    fn attrs(&self) -> &Vec<syn::Attribute> {
+        match self {
+            FnType::Fn(fun) => &fun.attrs,
+            #[cfg(feature = "poem")]
+            FnType::Method(fun) => &fun.attrs,
+        }
     }
 }
